requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-admin-thumbnails==0.2.8
django-debug-toolbar==4.2.0
django-filter==23.4
django-richtextfield==1.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
Faker==20.1.0
idna==3.6
inflection==0.5.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
uuid==1.30
